+++ app/core/consts.py
@@ -1,0 +1,4 @@
+from app.config import Config
+
+WEATHER_API_KEY = Config.WEATHER_API_KEY
+WEATHER_API_URL = "https://weather.visualcrossing.com/VisualCrossingWebServices/rest/services/timeline"
