+++ requirements.txt
@@ -2,3 +2,4 @@ fastapi
 uvicorn
 dotenv
 pydantic
+requests
